pyproject.toml
[tool.poetry]
name = "fastapi-signoz-test"
version = "0.1.0"
description = "FastAPI SigNoz Integration"
authors = ["Lalu Erfandi Maula Yusnu <nunenuh@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
openai = "^1.10.0"
pre-commit = "^4.0.1"
python-multipart = "^0.0.9"
pillow = "^10.2.0"
pytesseract = "^0.3.13"
pandas = "^2.2.3"
pdf2image = "^1.17.0"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-instrumentation-requests = "^0.43b0"
opentelemetry-exporter-otlp = "^1.22.0"
opentelemetry-instrumentation-logging = "^0.43b0"
opentelemetry-instrumentation-sqlalchemy = "^0.43b0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
pytest-cov = "^4.1.0"
autoflake = "^2.2.1"
jupyter = "^1.0.0"
jupyterlab = "^4.3.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
